refactor(logging): route server/MCP bypass sites through their structured loggers
Both server surfaces have a proper structured logger — `server::util::log`
(`{ts} {LVL} msg`, stdout + optional `--log-file`, CRLF-sanitized) and
`DalfoxMcp::log` (`[ts] [LVL] msg`, stderr — stdout is the JSON-RPC channel).
Yet several diagnostics bypassed them with ad-hoc `eprintln!("[server] …")` /
`eprintln!("[MCP][ERR] …")` lines: wrong stream, no timestamp/level, no
sanitization, and (server) absent from the log file.

Route every reachable bypass through the structured logger:
- server: graceful-shutdown notice + serve error (mod.rs); runtime-build-fail,
  task-panic, and recovery-runtime-fail in job_runner. All had `AppState`
  (or a clone) in scope. The shutdown future now takes a `state.clone()` since
  `state` is borrowed again for the serve-error log.
- MCP: thread-runtime build failure + scan-task panic now call
  `DalfoxMcp::log("ERR", …)` (stays on stderr — protocol-safe).

Also harden and de-duplicate the sanitizer:
- Move `sanitize_log_message` out of `server::util` into `crate::utils::log`
  (the shared logging module) so both servers depend on a neutral util rather
  than MCP reaching into `server`.
- `DalfoxMcp::log` now sanitizes too — it logs the same attacker-influenced
  bytes (scan URLs, panic payloads) the server logger already defends against.

Left as raw `eprintln!`: the pre-`AppState` startup/config errors (invalid
bind address, invalid allowed-origins regex/wildcard) — no logger exists yet.

Verified: `dalfox server` startup + SIGINT both emit `{ts} SERVER …` via the
logger; clippy clean; `cargo test --lib` (1756, single-threaded) passes.
